Constants

Cosmology.constants.G — Constant

Newton's gravitational constant in units of kpc km^2 / Msun / s^2.

Cosmology.constants.G_CGS — Constant

Newton's gravitational constant in units of cm^3 / g / s^2.

Cosmology.constants.KPC_CM — Constant

Number of centimeters in a kiloparsec.

Cosmology.constants.KPC_KM — Constant

Number of kilometers in a kiloparsec.

Cosmology.constants.DELTA_COLLAPSE — Constant

Linear overdensity threshold for halo collapse for the spherical top-hat collapse model, $3/5 \times (3π/2)^{2/3}$.

Cosmology.constants.RHO_C_Z0_CGS — Constant

Prefactor for the critical density of the Universe, in units of g / cm^3,

\[\frac{3}{8πG} \ \left(100 \ \text{km} \ \text{s}^{-1} \ \text{Mpc}^{-1} \right)^2\]

Cosmology.constants.RHO_C_Z0_KPC — Constant

Prefactor for the critical density of the Universe, in units of Msun / kpc^3,

\[\frac{3}{8πG} \ \left(100 \ \text{km} \ \text{s}^{-1} \ \text{Mpc}^{-1} \right)^2\]

Cosmology.constants.RHO_C_Z0_MPC — Constant

Prefactor for the critical density of the Universe, in units of Msun / Mpc^3,

\[\frac{3}{8πG} \ \left(100 \ \text{km} \ \text{s}^{-1} \ \text{Mpc}^{-1} \right)^2\]

Cosmology.constants.TNU_PREFAC — Constant

Conversion factor between CMB temperature and neutrino temperature, $\frac{4}{11}^{1/3}$.