API/Reference

Units

DynamicQuantities.jl and Unitful.jl work seamlessly with Cosmology.jl. In order to use their features, make sure to run either:

pkg> add DynamicQuantities
julia> using DynamicQuantities

or:

pkg> add Unitful UnitfulAstro
julia> using Unitful, UnitfulAstro

Cosmological Models

Cosmology.cosmology — Function

cosmology(; h = 0.69,
            Neff = 3.04,
            OmegaK = 0,
            OmegaM = 0.29,
            OmegaR = nothing,
            Tcmb = 2.7255,
            w0 = -1,
            wa = 0)

Parameters

h - Dimensionless Hubble constant

Neff - Effective number of massless neutrino species; used to compute Ω_ν

OmegaK - Curvature density (Ω_k)

OmegaM - Matter density (Ω_m)

OmegaR - Radiation density (Ω_r)

Tcmb - CMB temperature in Kelvin; used to compute Ω_γ

w0 - CPL dark energy equation of state; w = w0 + wa(1-a)

wa - CPL dark energy equation of state; w = w0 + wa(1-a)

Examples

julia> using Cosmology

julia> c = cosmology()
Cosmology.FlatLCDM{Float64}(0.69, 0.7099122024007928, 0.29, 8.77975992071536e-5)

julia> c = cosmology(OmegaK=0.1)
Cosmology.OpenLCDM{Float64}(0.69, 0.1, 0.6099122024007929, 0.29, 8.77975992071536e-5)

julia> c = cosmology(w0=-0.9, OmegaK=-0.1)
Cosmology.ClosedWCDM{Float64}(0.69, -0.1, 0.8099122024007929, 0.29, 8.77975992071536e-5, -0.9, 0.0)

Distances

Cosmology.angular_diameter_dist — Function

angular_diameter_dist([u::Unitlike,] c::AbstractCosmology, [z₁,] z₂)

Ratio of the proper transverse size in Mpc of an object at redshift z₂ to its angular size in radians, as seen by an observer at z₁. Redshift z₁ defaults to 0 if omitted. Will convert to compatible unit u if provided.

Cosmology.comoving_radial_dist — Function

comoving_radial_dist([u::Unitlike,] c::AbstractCosmology, [z₁,] z₂)

Comoving radial distance ($D_C$) in Mpc at redshift z₂ as seen by an observer at z₁. Redshift z₁ defaults to 0 if omitted. Will convert to compatible unit u if provided.

It's calculated as $D_C = D_{H0} Z$, where $D_{H0}$ is the Hubble distance at the present epoch and, $Z = \int_{z_1}^{z_2} \frac{dz}{E(z)}$.

Cosmology.luminosity_dist — Function

luminosity_dist([u::Unitlike,] c::AbstractCosmology, z)

Bolometric luminosity distance in Mpc at redshift z. Will convert to compatible unit u if provided.

Cosmology.distmod — Function

distmod(c::AbstractCosmology, z)

Distance modulus in magnitudes at redshift z.

Cosmology.hubble_dist — Function

hubble_dist(c::AbstractCosmology, z)

Hubble distance $D_H$, defined as the product of the speed of light and the Hubble time. That is, $D_H(z) = c / H(z)$.

See also

hubble_time

Examples

julia> c = cosmology(OmegaM=0.26)
Cosmology.FlatLCDM{Float64}(0.69, 0.7399122024007928, 0.26, 8.77975992071536e-5)

julia> angular_diameter_dist(c, 1.2)
1784.0089227105113 Mpc

julia> angular_diameter_dist(c, 0.7, 1.2)
606.6521737365097 Mpc

julia> luminosity_dist(c, 1.5)
11420.338287150073 Mpc

julia> luminosity_dist(us"Constants.Gpc", c, 1.5) # Can convert to appropriate unit
11.420338287150072 Gpc

Volumes

Cosmology.comoving_volume_element — Function

comoving_volume_element([u::Unitlike,] c::AbstractCosmology, z)

Comoving volume element in cubic Gpc out to redshift z. Will convert to compatible unit u if provided.

Cosmology.comoving_volume — Function

comoving_volume([u::Unitlike,] c::AbstractCosmology, z)

Comoving volume in cubic Gpc out to redshift z. Will convert to compatible unit u if provided.

Examples

julia> c = cosmology(OmegaM=0.26)
Cosmology.FlatLCDM{Float64}(0.69, 0.7399122024007928, 0.26, 8.77975992071536e-5)

julia> comoving_volume_element(c, 2.1)
46.74459228888613 Gpc³

julia> comoving_volume(c, 0.6)
49.363343663130685 Gpc³

julia> comoving_volume(us"Constants.ly^3", c, 0.6)
1.7127035381752996e30 ly³

Times

Cosmology.age — Function

age([u::Unitlike,] c::AbstractCosmology, z)

Age of the universe in Gyr at redshift z. Will convert to compatible unit u if provided.

Cosmology.lookback_time — Function

lookback_time([u::Unitlike,] c::AbstractCosmology, z)

Difference between age at redshift 0 and age at redshift z in Gyr. Will convert to compatible unit u if provided.

Cosmology.hubble_time — Function

hubble_time(c::AbstractCosmology, z)

Hubble time, defined as the inverse of the Hubble parameter. That is, $t_H(z) = 1/H(z)$.

See also

hubble_dist

Examples

julia> c = cosmology(OmegaM=0.26)
Cosmology.FlatLCDM{Float64}(0.69, 0.7399122024007928, 0.26, 8.77975992071536e-5)

julia> age(c, 1.2)
5.4454795007229455 Gyr

julia> lookback_time(us"yr", c, 1.2)
8.761465604385489e9 yr

Miscellaneous

Cosmology.H — Function

H(c::AbstractCosmology, z)

Hubble parameter at redshift z.

Cosmology.scale_factor — Function

scale_factor(z)

Return the scale factor $a(t)$ for a given redshift $z(t)$. According to the Friedmann–Lemaître–Robertson–Walker metric it's given as $a = 1/(1 + z)$ (Schneider 2015, p. 186).

A scale factor of 1, i.e., a redshift of 0, refers to the present epoch.

Bibliography

Schneider, P. (2015). Extragalactic Astronomy and Cosmology: An Introduction. 2nd Edition (Springer, Berlin, Heidelberg).